requirements.txt
fastapi
uvicorn
python-dotenv
supabase
mistralai
pydantic
python-jose


